+++ eslint.config.js
@@ -36,6 +36,7 @@ export default tseslint.config(
     languageOptions: {
       globals: {
         console: 'readonly',
+        global: 'readonly',
       },
     },
   },